+++ app/lib/ui/components/components.dart
@@ -1,11 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/ods_flutter_app_localizations.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/elevated_button.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/filled_button.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/filled_tonal_button.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/outlined_button.dart';
+import 'package:ods_flutter_demo/ui/components/buttons/text_button.dart';
 import 'package:ods_flutter_demo/ui/components/cards/card_small.dart';
 import 'package:ods_flutter_demo/ui/components/cards/card_vertical_image_first.dart';
 import 'package:ods_flutter_demo/ui/components/checkboxes/component_checkboxes.dart';
 import 'package:ods_flutter_demo/ui/components/component_entities.dart';
 import 'package:ods_flutter_demo/ui/components/material/component_material.dart';
-import 'package:ods_flutter_demo/ui/components/navigation_bar/navigation_bar.dart';
 import 'package:ods_flutter_demo/ui/components/radio_buttons/component_radio_buttons.dart';
 import 'package:ods_flutter_demo/ui/components/switches/component_switches.dart';
 
@@ -60,14 +64,30 @@ List<Component> components(BuildContext context) {
       ],
     ),
     Component(
-      AppLocalizations.of(context)!.componentNavigationBarTitle,
-      'assets/il_navigation_bar.png',
-      AppLocalizations.of(context)!.componentNavigationBarDescription,
+      AppLocalizations.of(context)!.componentButtonsTitle,
+      'assets/il_buttons.svg',
+      AppLocalizations.of(context)!.componentButtonsDescription,
       [
         Variant(
-            AppLocalizations.of(context)!.navigationBarVariantTitle,
-            AppLocalizations.of(context)!.navigationBarVariantSubtitle,
-            ComponentNavigationBar()),
+            AppLocalizations.of(context)!.elevatedButtonsVariantTitle,
+            AppLocalizations.of(context)!.elevatedButtonsVariantSubtitle,
+            ComponentElevatedButtons()),
+        Variant(
+            AppLocalizations.of(context)!.filledButtonsVariantTitle,
+            AppLocalizations.of(context)!.filledButtonsVariantSubtitle,
+            ComponentFilledButtons()),
+        Variant(
+            AppLocalizations.of(context)!.filledTonalButtonsVariantTitle,
+            AppLocalizations.of(context)!.filledTonalButtonsVariantSubtitle,
+            ComponentFilledTonalButtons()),
+        Variant(
+            AppLocalizations.of(context)!.outlinedButtonsVariantTitle,
+            AppLocalizations.of(context)!.outlinedButtonsVariantSubtitle,
+            ComponentOutlinedButtons()),
+        Variant(
+            AppLocalizations.of(context)!.textButtonsVariantTitle,
+            AppLocalizations.of(context)!.textButtonsVariantSubtitle,
+            ComponentTextButtons()),
       ],
     ),
     Component(