+++ app/CHANGELOG.md
@@ -15,8 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Review the string in the app to use localization keys ([#129](https://github.com/Orange-OpenSource/ods-flutter/issues/129))
 - Add `OdsVerticalImageFirstCard` widget demonstration ([#8](https://github.com/Orange-OpenSource/ods-flutter/issues/8))
 - Add recipes Json file and use it ([#123](https://github.com/Orange-OpenSource/ods-flutter/issues/123))
-- Add customization bottom sheet to configure components ([#148](https://github.com/Orange-OpenSource/ods-flutter/issues/148)))
+- Add customization bottom sheet to configure components ([#148](https://github.com/Orange-OpenSource/ods-flutter/issues/148))
 - Add component navigation bar ([#23](https://github.com/Orange-OpenSource/ods-flutter/issues/23))
+- Create different screens for the variants of button ([#105](https://github.com/Orange-OpenSource/ods-flutter/issues/105))
 
 ### Changed